backend/requirements.txt
azure-functions>=1.11.0,<2.0.0
azure-cosmos>=4.3.0
requests>=2.28.0